# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Blogger API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthUrl {
  url: string;
}

export interface AuthStatus {
  authenticated: boolean;
  blogTitle?: string;
  blogUrl?: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export type PostStatus = (typeof PostStatus)[keyof typeof PostStatus];

export const PostStatus = {
  published: "published",
  draft: "draft",
  failed: "failed",
} as const;

export interface Post {
  id: number;
  title: string;
  topic: string;
  niche: string;
  status: PostStatus;
  bloggerPostId?: string;
  bloggerUrl?: string;
  labels: string[];
  metaDescription?: string;
  wordCount: number;
  createdAt: string;
  publishedAt?: string;
}

export interface PostList {
  posts: Post[];
  total: number;
}

export type GeneratePostBodyNiche =
  (typeof GeneratePostBodyNiche)[keyof typeof GeneratePostBodyNiche];

export const GeneratePostBodyNiche = {
  motivation: "motivation",
  ai: "ai",
  money: "money",
  facts: "facts",
  tech: "tech",
} as const;

export interface GeneratePostBody {
  topic?: string;
  niche?: GeneratePostBodyNiche;
}

export interface GeneratePostResult {
  success: boolean;
  post?: Post;
  message?: string;
}

export type PostStatsByNicheItem = {
  niche: string;
  count: number;
};

export interface PostStats {
  totalPublished: number;
  totalFailed: number;
  totalDraft: number;
  todayCount: number;
  thisWeekCount: number;
  byNiche: PostStatsByNicheItem[];
  recentTitles: string[];
}

export interface SchedulerStatus {
  enabled: boolean;
  postsPerDay: number;
  nextRunAt?: string;
  lastRunAt?: string;
  isRunning: boolean;
}

export type LogLevel = (typeof LogLevel)[keyof typeof LogLevel];

export const LogLevel = {
  info: "info",
  error: "error",
  warn: "warn",
} as const;

export interface Log {
  id: number;
  level: LogLevel;
  message: string;
  details?: string;
  createdAt: string;
}

export interface LogList {
  logs: Log[];
  total: number;
}

export interface Settings {
  blogId: string;
  postsPerDay: number;
  schedulerEnabled: boolean;
  niches: string[];
  pexelsApiKey?: string;
  imagesPerPost: number;
}

export interface UpdateSettingsBody {
  postsPerDay?: number;
  schedulerEnabled?: boolean;
  niches?: string[];
  pexelsApiKey?: string;
  imagesPerPost?: number;
}

export type HandleAuthCallbackParams = {
  code?: string;
  error?: string;
};

export type ListPostsParams = {
  limit?: number;
  offset?: number;
  status?: ListPostsStatus;
};

export type ListPostsStatus =
  (typeof ListPostsStatus)[keyof typeof ListPostsStatus];

export const ListPostsStatus = {
  published: "published",
  draft: "draft",
  failed: "failed",
  all: "all",
} as const;

export type ListLogsParams = {
  limit?: number;
  level?: ListLogsLevel;
};

export type ListLogsLevel = (typeof ListLogsLevel)[keyof typeof ListLogsLevel];

export const ListLogsLevel = {
  info: "info",
  error: "error",
  warn: "warn",
  all: "all",
} as const;