# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AI Blogger API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /posts:
    get:
      operationId: listPosts
      tags: [posts]
      summary: List all blog posts
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
        - name: offset
          in: query
          required: false
          schema:
            type: integer
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [published, draft, failed, all]
      responses:
        "200":
          description: List of posts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostList"
  /posts/generate:
    post:
      operationId: generatePost
      tags: [posts]
      summary: Generate and publish a new blog post
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GeneratePostBody"
      responses:
        "200":
          description: Generated post result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneratePostResult"
  /posts/{id}:
    get:
      operationId: getPost
      tags: [posts]
      summary: Get a blog post by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Post details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
...
components:
  schemas:
...
    Post:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        topic:
          type: string
        niche:
          type: string
        status:
          type: string
          enum: [published, draft, failed]
        bloggerPostId:
          type: string
        bloggerUrl:
          type: string
        labels:
          type: array
          items:
            type: string
        metaDescription:
          type: string
        wordCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        publishedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - topic
        - niche
        - status
        - labels
        - wordCount
        - createdAt
    PostList:
      type: object
      properties:
        posts:
          type: array
          items:
            $ref: "#/components/schemas/Post"
        total:
          type: integer
      required:
        - posts
        - total
    GeneratePostBody:
      type: object
      properties:
        topic:
          type: string
        niche:
          type: string
          enum: [motivation, ai, money, facts, tech]
    GeneratePostResult:
      type: object
      properties:
        success:
          type: boolean
        post:
          $ref: "#/components/schemas/Post"
        message:
          type: string
      required:
        - success
...