# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Blogger API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Get Google OAuth URL
 */
export const GetAuthUrlResponse = zod.object({
  url: zod.string(),
});

/**
 * @summary Get authentication status
 */
export const GetAuthStatusResponse = zod.object({
  authenticated: zod.boolean(),
  blogTitle: zod.string().optional(),
  blogUrl: zod.string().optional(),
});

/**
 * @summary Handle OAuth callback
 */
export const HandleAuthCallbackQueryParams = zod.object({
  code: zod.coerce.string().optional(),
  error: zod.coerce.string().optional(),
});

/**
 * @summary Logout and clear tokens
 */
export const LogoutResponse = zod.object({
  success: zod.boolean(),
  message: zod.string().optional(),
});

/**
 * @summary List all blog posts
 */
export const ListPostsQueryParams = zod.object({
  limit: zod.coerce.number().optional(),
  offset: zod.coerce.number().optional(),
  status: zod.enum(["published", "draft", "failed", "all"]).optional(),
});

export const ListPostsResponse = zod.object({
  posts: zod.array(
    zod.object({
      id: zod.number(),
      title: zod.string(),
      topic: zod.string(),
      niche: zod.string(),
      status: zod.enum(["published", "draft", "failed"]),
      bloggerPostId: zod.string().optional(),
      bloggerUrl: zod.string().optional(),
      labels: zod.array(zod.string()),
      metaDescription: zod.string().optional(),
      wordCount: zod.number(),
      createdAt: zod.coerce.date(),
      publishedAt: zod.coerce.date().optional(),
    }),
  ),
  total: zod.number(),
});

/**
 * @summary Generate and publish a new blog post
 */
export const GeneratePostBody = zod.object({
  topic: zod.string().optional(),
  niche: zod.enum(["motivation", "ai", "money", "facts", "tech"]).optional(),
});

export const GeneratePostResponse = zod.object({
  success: zod.boolean(),
  post: zod
    .object({
      id: zod.number(),
      title: zod.string(),
      topic: zod.string(),
      niche: zod.string(),
      status: zod.enum(["published", "draft", "failed"]),
      bloggerPostId: zod.string().optional(),
      bloggerUrl: zod.string().optional(),
      labels: zod.array(zod.string()),
      metaDescription: zod.string().optional(),
      wordCount: zod.number(),
      createdAt: zod.coerce.date(),
      publishedAt: zod.coerce.date().optional(),
    })
    .optional(),
  message: zod.string().optional(),
});

/**
 * @summary Get a blog post by ID
 */
export const GetPostParams = zod.object({
  id: zod.coerce.number(),
});

export const GetPostResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  topic: zod.string(),
  niche: zod.string(),
  status: zod.enum(["published", "draft", "failed"]),
  bloggerPostId: zod.string().optional(),
  bloggerUrl: zod.string().optional(),
  labels: zod.array(zod.string()),
  metaDescription: zod.string().optional(),
  wordCount: zod.number(),
  createdAt: zod.coerce.date(),
  publishedAt: zod.coerce.date().optional(),
});

/**
 * @summary Get posting statistics
 */
export const GetPostStatsResponse = zod.object({
  totalPublished: zod.number(),
  totalFailed: zod.number(),
  totalDraft: zod.number(),
  todayCount: zod.number(),
  thisWeekCount: zod.number(),
  byNiche: zod.array(
    zod.object({
      niche: zod.string(),
      count: zod.number(),
    }),
  ),
  recentTitles: zod.array(zod.string()),
});

/**
 * @summary Get scheduler status and next run time
 */
export const GetSchedulerStatusResponse = zod.object({
  enabled: zod.boolean(),
  postsPerDay: zod.number(),
  nextRunAt: zod.coerce.date().optional(),
  lastRunAt: zod.coerce.date().optional(),
  isRunning: zod.boolean(),
});

/**
 * @summary Manually trigger the scheduler to run now
 */
export const RunSchedulerNowResponse = zod.object({
  success: zod.boolean(),
  message: zod.string().optional(),
});

/**
 * @summary List activity logs
 */
export const ListLogsQueryParams = zod.object({
  limit: zod.coerce.number().optional(),
  level: zod.enum(["info", "error", "warn", "all"]).optional(),
});

export const ListLogsResponse = zod.object({
  logs: zod.array(
    zod.object({
      id: zod.number(),
      level: zod.enum(["info", "error", "warn"]),
      message: zod.string(),
      details: zod.string().optional(),
      createdAt: zod.coerce.date(),
    }),
  ),
  total: zod.number(),
});

/**
 * @summary Get current settings
 */
export const GetSettingsResponse = zod.object({
  blogId: zod.string(),
  postsPerDay: zod.number(),
  schedulerEnabled: zod.boolean(),
  niches: zod.array(zod.string()),
  pexelsApiKey: zod.string().optional(),
  imagesPerPost: zod.number(),
});

/**
 * @summary Update settings
 */
export const UpdateSettingsBody = zod.object({
  postsPerDay: zod.number().optional(),
  schedulerEnabled: zod.boolean().optional(),
  niches: zod.array(zod.string()).optional(),
  pexelsApiKey: zod.string().optional(),
  imagesPerPost: zod.number().optional(),
});

export const UpdateSettingsResponse = zod.object({
  blogId: zod.string(),
  postsPerDay: zod.number(),
  schedulerEnabled: zod.boolean(),
  niches: zod.array(zod.string()),
  pexelsApiKey: zod.string().optional(),
  imagesPerPost: zod.number(),
});