# lib/api-zod/src/generated/types/generatePostBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Blogger API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GeneratePostBodyNiche } from "./generatePostBodyNiche";

export interface GeneratePostBody {
  topic?: string;
  niche?: GeneratePostBodyNiche;
}