# lib/api-zod/src/generated/types/postList.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Blogger API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Post } from "./post";

export interface PostList {
  posts: Post[];
  total: number;
}